IMP: Multi platform and multi architecture support improvement. ADD: Versioning mechanisms added to cmake scripts. Now it automatically gets revision number from svn. ADD: Version.txt added.Version.txt should include version of the source. IMP: Platform checks added to cmake functions ze_add_xxxxx functions. ADD: New ze_add_module command added to cmake functions. REM: ze_add_conditional_directorty function removed from cmake functions. Instead of ze_add_conditional_directory use use ze_add_module. CHG: Lib, Include and Dll folders moved to Platform folder in [Platform Name]\[Architecture] directory structure. CHG: All sub-system modules moved to Source/ZEModules directory. CHG: ZETest library moved to ZEFoundation.
svn path=/trunk/; revision=845

+++ Source/ZEEngine/ZECore/ZEProfilerCounter.cpp
@@ -74,49 +74,54 @@ const ZEString& ZEProfilerCounter::GetName()
 void ZEProfilerCounter::Start()
 {
 	CounterActive = true;
-	StartTime = 0;
 	EndTime = 0;
-	StartTime = GetTickCount64();
+
+	LARGE_INTEGER Temp;
+	QueryPerformanceCounter(&Temp);
+
+	StartTime = Temp.QuadPart;
 }
 
 void ZEProfilerCounter::Stop()
 {
-	EndTime = GetTickCount64();
-	if(Owner->GetInstance()->Stack.GetLastItem()->Name == ZEProfilerCounter::Name)
-	{
-		PassedTime = EndTime - StartTime;
+	LARGE_INTEGER Temp;
+	QueryPerformanceCounter(&Temp);
 
-		if(FrameMaximumTime < 0)
-			FrameMaximumTime = PassedTime;
-		else if(PassedTime > FrameMaximumTime)
-			FrameMaximumTime = PassedTime;
+	EndTime = Temp.QuadPart;
 
-		if(FrameMinimumTime < 0)
-			FrameMinimumTime = PassedTime;
-		else if(FrameMinimumTime > PassedTime)
-			FrameMinimumTime = PassedTime;
+	PassedTime = EndTime - StartTime;
 
-		if(TotalMaximumTime < 0)
-			TotalMaximumTime = PassedTime;
-		else if(TotalMaximumTime < PassedTime)
-			TotalMaximumTime = PassedTime;
+	if(FrameMaximumTime < 0)
+		FrameMaximumTime = PassedTime;
+	else if(PassedTime > FrameMaximumTime)
+		FrameMaximumTime = PassedTime;
 
-		if(TotalMinimumTime < 0)
-			TotalMinimumTime = PassedTime;
-		else if(TotalMinimumTime > PassedTime)
-			TotalMinimumTime = PassedTime;
+	if(FrameMinimumTime < 0)
+		FrameMinimumTime = PassedTime;
+	else if(FrameMinimumTime > PassedTime)
+		FrameMinimumTime = PassedTime;
 
-		FrameTotalTime += PassedTime;
-		TotalTime += PassedTime;
-		TotalCount++;
-		FrameCount++;
+	if(TotalMaximumTime < 0)
+		TotalMaximumTime = PassedTime;
+	else if(TotalMaximumTime < PassedTime)
+		TotalMaximumTime = PassedTime;
 
-		CounterActive = false;
-		StartTime = 0;
-		EndTime = 0;
+	if(TotalMinimumTime < 0)
+		TotalMinimumTime = PassedTime;
+	else if(TotalMinimumTime > PassedTime)
+		TotalMinimumTime = PassedTime;
+
+	FrameTotalTime += PassedTime;
+	TotalTime += PassedTime;
+	TotalCount++;
+	FrameCount++;
+
+	CounterActive = false;
+	StartTime = 0;
+	EndTime = 0;
+		
+	if(Owner != NULL && Owner->GetInstance()->Stack.GetLastItem()->Name == ZEProfilerCounter::Name)
 		Owner->GetInstance()->Stack.DeleteAt(Owner->GetInstance()->Stack.GetSize()-1);
-	}
-	
 }
 
 ZEInt64 ZEProfilerCounter::GetFrameCount()
